+++ pytorch_lightning/accelerators/accelerator_connector.py
@@ -7,8 +7,8 @@ from pytorch_lightning.utilities import rank_zero_only
 from pytorch_lightning.utilities.distributed import rank_zero_warn, rank_zero_info
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning import _logger as log
-from pytorch_lightning.cluster_environments.slurm_environment import SLURMEnvironment
-from pytorch_lightning.cluster_environments.torchelastic_environment import TorchElasticEnvironment
+from pytorch_lightning.environments.slurm_environment import SLURMEnvironment
+from pytorch_lightning.environments.torchelastic_environment import TorchElasticEnvironment
 
 try:
     import torch_xla
