--- pytorch_lightning/cluster_environments/__init__.py
@@ -1,3 +1,0 @@
-from pytorch_lightning.cluster_environments.cluster_environment import ClusterEnvironment
-from pytorch_lightning.cluster_environments.slurm_environment import SLURMEnvironment
-from pytorch_lightning.cluster_environments.torchelastic_environment import TorchElasticEnvironment
