+++ pytorch_lightning/environments/slurm_environment.py
@@ -1,7 +1,7 @@
 import os
 import re
 from pytorch_lightning import _logger as log
-from pytorch_lightning.cluster_environments.cluster_environment import ClusterEnvironment
+from pytorch_lightning.environments.cluster_environment import ClusterEnvironment
 
 
 class SLURMEnvironment(ClusterEnvironment):
