+++ pytorch_lightning/environments/torchelastic_environment.py
@@ -1,7 +1,7 @@
 import os
 from pytorch_lightning import _logger as log
 from pytorch_lightning.utilities import rank_zero_warn
-from pytorch_lightning.cluster_environments.cluster_environment import ClusterEnvironment
+from pytorch_lightning.environments.cluster_environment import ClusterEnvironment
 
 
 class TorchElasticEnvironment(ClusterEnvironment):
